refactor: reusable `NftDataForm` component
