+++ src/components/[guild]/RolePlatforms/components/AddRoleRewardModal/components/AddContractCallPanel/components/CreateNftForm/components/ImagePicker.tsx
@@ -5,40 +5,46 @@ import {
   FormLabel,
   Icon,
   Img,
+  Progress,
   Text,
   VStack,
 } from "@chakra-ui/react"
 import Button from "components/common/Button"
 import FormErrorMessage from "components/common/FormErrorMessage"
 import useDropzone from "hooks/useDropzone"
+import usePinata from "hooks/usePinata"
 import { Image } from "phosphor-react"
 import { useState } from "react"
 import { useController, useFormContext } from "react-hook-form"
-import { CreateNftFormType } from "../CreateNftForm"
+import { CreateNftFormType } from "./NftDataForm"
 
 const ImagePicker = () => {
   const {
-    setValue,
+    control,
     formState: { errors },
   } = useFormContext<CreateNftFormType>()
 
-  useController({
+  const { field: imageField } = useController({
     name: "image",
     rules: {
       required: "It is required to upload an image",
     },
   })
 
-  const [preview, setPreview] = useState<string>()
-
+  const { isUploading, onUpload } = usePinata({
+    control,
+    fieldToSetOnSuccess: "image",
+  })
+  const [progress, setProgress] = useState(0)
   const { isDragActive, fileRejections, getRootProps, getInputProps } = useDropzone({
     multiple: false,
     noClick: false,
     maxSizeMb: 100,
     onDrop: (acceptedFiles) => {
-      setPreview(URL.createObjectURL(acceptedFiles[0]))
-      setValue("image", acceptedFiles[0], {
-        shouldValidate: true,
+      if (!acceptedFiles[0]) return
+      onUpload({
+        data: [acceptedFiles[0]],
+        onProgress: setProgress,
       })
     },
   })
@@ -48,9 +54,37 @@ const ImagePicker = () => {
       <FormLabel>Media</FormLabel>
 
       <AspectRatio ratio={1}>
-        <Button p={0} borderWidth={1} variant="ghost" {...getRootProps()}>
-          {preview ? (
-            <Img aspectRatio={1} objectFit="cover" src={preview} alt="NFT image" />
+        <Button
+          p={0}
+          borderWidth={1}
+          variant="ghost"
+          w="full"
+          {...getRootProps()}
+          sx={{
+            "> div": {
+              width: "100%",
+            },
+          }}
+        >
+          {isUploading ? (
+            <VStack p={4}>
+              <Text colorScheme="gray" fontSize="sm" fontWeight="normal">
+                Uploading...
+              </Text>
+              <Progress
+                borderRadius="full"
+                w="full"
+                value={progress * 100}
+                colorScheme="gray"
+              />
+            </VStack>
+          ) : imageField.value ? (
+            <Img
+              aspectRatio={1}
+              objectFit="cover"
+              src={imageField.value}
+              alt="NFT image"
+            />
           ) : (
             <VStack p={4}>
               <Icon as={Image} weight="light" boxSize={16} color="gray" />
